Move /api/token endpoint into application
This starts the process of moving all auth-related code out of the
`h.api` package and into the application. Token issuance is the domain
of the web application, as it will eventually depend on more than just
the current authenticated session.
